Editor Icons shipped with Build
Updated texture usage in ImGui
Asset serialization rework
Asset Inspector
Asset parameter exposure functionality
Add functionality to modify and save Asset files
DirectX shader compilation functionality
Engine shader composition design (HLSL Code + Shader Config = Engine Shader)
Added grouping support for exposed variables

+++ GameSmithEngine/src/GameSmithEngine/SerializeableFiles/GenericFileTypes/JsonFile.cpp
@@ -20,7 +20,7 @@ namespace GameSmith {
 
 	unsigned int JsonFile::RequiredSpace() const
 	{
-		return m_JsonString.length() + 1;
+		return ((unsigned int)m_JsonString.length()) + 1;
 	}
 
 	void JsonFile::Deserialize(char* inData, unsigned int size)
